fix bug with int, replace with uint64_t
Former-commit-id: adee864807e2c06dca05ce394dbba0745ca15751

+++ src/kmer-index/bitarray.c
@@ -2,7 +2,7 @@
 #include<stdlib.h>
 #include "bitarray.h"
 
-bitarray_t* create_bitarray(int n)
+bitarray_t* create_bitarray(uint64_t n)
 {
   bitarray_t* array = malloc(sizeof(bitarray_t));
 	array->values = calloc(n, sizeof(char));
